move files & cleanup
